GPC-61: Refactor GPPRestClient
